Some PHPStan fixes

+++ src/DataType/CharDataType.php
@@ -8,6 +8,6 @@ class CharDataType extends AbstractDataType
 {
     public function populate(Column $column): string
     {
-        return $this->faker->realText(mt_rand(10, max($column->getLength(), 10)));
+        return $this->faker->realText(mt_rand(10, (int)max($column->getLength(), 10)));
     }
 }
