+++ src/DataType/TextDataType.php
@@ -6,12 +6,14 @@ use Populator\Structure\Column;
 
 class TextDataType extends AbstractDataType
 {
+    /** @var int */
     protected $min = 0;
 
+    /** @var int */
     protected $max = 65535;
 
     public function populate(Column $column): string
     {
-        return $this->faker->realText(mt_rand(max($this->min, 10), $this->max));
+        return $this->faker->realText(mt_rand((int)max($this->min, 10), $this->max));
     }
 }
